refactor: Main project의 package 이름 변경   - com.example.base -> com.github.can019
