--- src/main/java/com/example/base/BaseApplication.java
@@ -1,17 +1,0 @@
-package com.example.base;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
-import org.springframework.boot.web.servlet.ServletComponentScan;
-
-@SpringBootApplication
-@ConfigurationPropertiesScan("com.example.base.config")
-@ServletComponentScan
-public class BaseApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BaseApplication.class, args);
-	}
-
-}
