--- src/main/java/com/example/base/config/WebMvcConfig.java
@@ -1,16 +1,0 @@
-package com.example.base.config;
-
-import com.example.base.global.interceptor.LoggingInterceptor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebMvcConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(new LoggingInterceptor())
-                .excludePathPatterns("/css/**", "/images/**", "/js/**");
-    }
-}
