--- src/main/java/com/example/base/global/entity/BaseTimeEntity.java
@@ -1,20 +1,0 @@
-package com.example.base.global.entity;
-
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.time.LocalDateTime;
-
-@Getter
-@MappedSuperclass
-public abstract class BaseTimeEntity {
-    @CreationTimestamp
-    private LocalDateTime createdTime;
-
-    @UpdateTimestamp
-    private LocalDateTime updatedTime;
-
-
-}
