+++ src/main/java/com/github/can019/config/WebMvcConfig.java
@@ -1,0 +1,16 @@
+package com.github.can019.config;
+
+import com.github.can019.global.interceptor.LoggingInterceptor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebMvcConfig implements WebMvcConfigurer {
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(new LoggingInterceptor())
+                .excludePathPatterns("/css/**", "/images/**", "/js/**");
+    }
+}
