+++ src/main/java/com/github/can019/domain/user/domain/UUIDV1IdGenerator.java
@@ -1,0 +1,14 @@
+package com.github.can019.domain.user.domain;
+
+import com.github.can019.global.util.generator.UUID;
+import org.hibernate.engine.spi.SharedSessionContractImplementor;
+import org.hibernate.id.Configurable;
+import org.hibernate.id.IdentifierGenerator;
+
+public class UUIDV1IdGenerator implements IdentifierGenerator, Configurable {
+    private String method;
+    @Override
+    public Object generate(SharedSessionContractImplementor session, Object object) {
+     return UUID.generateSequentialUUIDV1WithoutHyphenAsByte();
+    }
+}
