+++ src/test/integration/java/com/example/base/domain/user/service/UserServiceTest.java
@@ -1,8 +1,9 @@
 package com.example.base.domain.user.service;
 
-import com.example.base.domain.user.domain.User;
-import com.example.base.domain.user.repository.UserRepository;
-import com.example.base.global.util.convertor.TypeConvertor;
+import com.github.can019.domain.user.domain.User;
+import com.github.can019.domain.user.repository.UserRepository;
+import com.github.can019.domain.user.service.UserService;
+import com.github.can019.global.util.convertor.TypeConvertor;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import jakarta.transaction.Transactional;
