+++ src/test/unit/java/com/example/base/global/convertor/TypeConvertorTest.java
@@ -1,7 +1,6 @@
 package com.example.base.global.convertor;
 
-import com.example.base.global.util.convertor.TypeConvertor;
-import org.assertj.core.api.Assertions;
+import com.github.can019.global.util.convertor.TypeConvertor;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
